+++ main/global_state.h
@@ -128,7 +128,7 @@ typedef struct
     uint8_t * valid_jobs;
     pthread_mutex_t valid_jobs_lock;
 
-    uint32_t stratum_difficulty;
+    double stratum_difficulty;
     uint32_t version_mask;
     bool new_stratum_version_rolling_msg;
